测试一下

+++ gulimall-product/src/main/java/com/atguigu/gulimall/product/entity/SkuInfoEntity.java
@@ -60,7 +60,7 @@ public class SkuInfoEntity implements Serializable {
 	/**
 	 * 价格
 	 */
-	private BigDecimal price;
+	private Long price;
 	/**
 	 * 销量
 	 */
